googlesearch/requirements.txt
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
google-genai>=1.0.0
openai>=1.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
